Updates include:     - (edit) rename Download/DownloadStart to DownloadStartProcessing     - (edit) move Download/[DownloadErrorDownloading, DownloadExplore, DownloadProcessing, DownloadReady, DownloadRegenerate] to Dataset/ and rename to Dataset/[DatasetErrorDownloading, DatasetExplore, DatasetProcessing, DatasetReady, DatasetRegenerate] respectively     - (delete) remove Download/DownloadFooterCTA     - create Dataset/[MoveToDatasetButton, ShareDatasetButton] and import to pages/dataset/[dataset_id]     - (temp) create a temp directory to render my dataset download views temporarily     - (minor) fix spacing for responsive     - (minor) add ‘/‘ to image paths in JSX(Grommet Box props) to prevent 404 when loading images     - (minor) add additional href attribute in Download/DownloadEmailForm     - (minor) remove the outer <Box> and unused imports from pages/download
