+++ src/components/Download/DownloadEmailForm.js
@@ -1,20 +1,17 @@
 import { useResponsive } from 'hooks/useResponsive'
-import { Box, CheckBox, Text } from 'grommet'
+import { Box, CheckBox, Form, Text } from 'grommet'
 import { Anchor } from 'components/shared/Anchor'
 import { Button } from 'components/shared/Button'
 import { Column } from 'components/shared/Column'
 import { Row } from 'components/shared/Row'
 import { TextInput } from 'components/shared/TextInput'
-
 import { links } from 'config'
 
-export const DownloadEmailForm = () => {
+export const DownloadEmailForm = ({ handleSubmit }) => {
   const { setResponsive } = useResponsive()
 
-  const handleClick = () => {}
-
   return (
-    <Box>
+    <Form onSubmit={handleSubmit}>
       <Row width="500px">
         <Column fill>
           <TextInput placeholder="jdoe@example.com" type="email" />
@@ -27,6 +24,7 @@ export const DownloadEmailForm = () => {
           }}
           primary
           responsive
+          type="submit"
         />
       </Row>
       <Box margin={{ top: 'small' }}>
@@ -34,16 +32,20 @@ export const DownloadEmailForm = () => {
           label={
             <Text>
               I agree to the{' '}
-              <Anchor href={links.terms} label="Terms of Use  " />
+              <Anchor
+                href={links.terms}
+                label="Terms of Use"
+                rel="noopener noreferrer"
+                target="_blank"
+              />
             </Text>
           }
-          onClick={handleClick}
         />
         <CheckBox
+          checked
           label="I would like to receive occasional updates from the refine.bio team"
-          onClick={handleClick}
         />
       </Box>
-    </Box>
+    </Form>
   )
 }
