+++ src/components/Download/index.js
@@ -1,6 +1,2 @@
 export { DownloadEmpty } from './DownloadEmpty'
-export { DownloadError } from './DownloadError'
-export { DownloadProcessing } from './DownloadProcessing'
-export { DownloadReady } from './DownloadReady'
-export { DownloadRegenerate } from './DownloadRegenerate'
-export { DownloadStart } from './DownloadStart'
+export { DownloadStartProcessing } from './DownloadStartProcessing'
